Move src/simd to Platforms/CPU/SIMD.
