+++ src/Numerics/MatrixOperators.h
@@ -22,7 +22,7 @@
 #include "OhmmsPETE/TinyVector.h"
 #include "OhmmsPETE/Tensor.h"
 #include "Numerics/OhmmsBlas.h"
-#include "simd/inner_product.hpp"
+#include "CPU/SIMD/inner_product.hpp"
 
 namespace qmcplusplus
 {
