+++ src/Numerics/tests/CMakeLists.txt
@@ -20,7 +20,7 @@ SET(UTEST_NAME deterministic-unit_test_${SRC_DIR})
 
 
 SET(UTEST_SRCS test_grid_functor.cpp test_stdlib.cpp test_bessel.cpp
-               test_ylm.cpp test_e2iphi.cpp test_aligned_allocator.cpp
+               test_ylm.cpp test_e2iphi.cpp
                test_gaussian_basis.cpp test_cartesian_tensor.cpp test_soa_cartesian_tensor.cpp
                test_transform.cpp test_min_oned.cpp test_one_dim_cubic_spline.cpp)
