+++ src/OhmmsSoA/Container.h
@@ -18,7 +18,6 @@
 #include <config.h>
 #include <OhmmsPETE/TinyVector.h>
 #include <OhmmsPETE/Tensor.h>
-#include <simd/allocator.hpp>
 #include <OhmmsSoA/PosTransformer.h>
 #include <OhmmsSoA/VectorSoaContainer.h>
 //#include <OhmmsSoA/TensorSoaContainer.h>
