+++ src/Platforms/CMakeLists.txt
@@ -26,6 +26,9 @@ ADD_LIBRARY(platform_device ${DEVICE_SRCS})
 TARGET_INCLUDE_DIRECTORIES(platform_device PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}")
 TARGET_LINK_LIBRARIES(platform_device PUBLIC message)
 
+SUBDIRS(CPU)
+TARGET_LINK_LIBRARIES(platform_device PRIVATE platform_cpu)
+
 SUBDIRS(OpenMP)
 TARGET_LINK_LIBRARIES(platform_device PRIVATE platform_omp)
