+++ src/Platforms/CPU/SIMD/algorithm.hpp
