+++ src/Platforms/CPU/SIMD/inner_product.hpp
