+++ src/Platforms/CPU/SIMD/vmath.hpp
