+++ src/Platforms/CPU/tests/CMakeLists.txt
@@ -2,23 +2,21 @@
 #// This file is distributed under the University of Illinois/NCSA Open Source License.
 #// See LICENSE file in top directory for details.
 #//
-#// Copyright (c) 2016 Jeongnim Kim and QMCPACK developers.
+#// Copyright (c) 2019 QMCPACK developers.
 #//
 #// File developed by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//
-#// File created by: Mark Dewing, markdewing@gmail.com, University of Illinois at Urbana-Champaign
+#// File created by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//////////////////////////////////////////////////////////////////////////////////////
-
 
 
 SET(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${QMCPACK_UNIT_TEST_DIR})
 
-SET(SRC_DIR ohmmspete)
+SET(SRC_DIR cpu)
 SET(UTEST_EXE test_${SRC_DIR})
 SET(UTEST_NAME deterministic-unit_test_${SRC_DIR})
 
-
-ADD_EXECUTABLE(${UTEST_EXE} test_vector.cpp test_matrix.cpp test_tinyvector.cpp)
-TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main containers)
+ADD_EXECUTABLE(${UTEST_EXE} test_aligned_allocator.cpp)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} platform_cpu catch_main)
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
