+++ src/QMCWaveFunctions/AGPDeterminant.cpp
@@ -15,7 +15,7 @@
 #include "QMCWaveFunctions/AGPDeterminant.h"
 #include "Numerics/DeterminantOperators.h"
 #include "Numerics/MatrixOperators.h"
-#include "simd/simd.hpp"
+#include "CPU/SIMD/simd.hpp"
 
 namespace qmcplusplus
 {
