+++ src/QMCWaveFunctions/Fermion/MatrixDelayedUpdateCUDA.h
@@ -12,7 +12,7 @@
 #ifndef QMCPLUSPLUS_MATRIX_DELAYED_UPDATE_CUDA_H
 #define QMCPLUSPLUS_MATRIX_DELAYED_UPDATE_CUDA_H
 
-#include "simd/allocator.hpp"
+#include "CPU/SIMD/aligned_allocator.hpp"
 #include "Platforms/PinnedAllocator.h"
 #include "OpenMP/OMPallocator.hpp"
 #include "OhmmsPETE/OhmmsVector.h"
