+++ src/QMCWaveFunctions/Jastrow/BsplineFunctor.h
@@ -24,7 +24,7 @@
 #include "Utilities/ProgressReportEngine.h"
 #include "OhmmsData/AttributeSet.h"
 #include "Numerics/LinearFit.h"
-#include "simd/allocator.hpp"
+#include "CPU/SIMD/aligned_allocator.hpp"
 #include <cstdio>
 
 namespace qmcplusplus
