+++ src/QMCWaveFunctions/tests/test_dirac_matrix.cpp
@@ -17,7 +17,7 @@
 #include "QMCWaveFunctions/WaveFunctionComponent.h"
 #include "QMCWaveFunctions/Fermion/DiracMatrix.h"
 #include "QMCWaveFunctions/Fermion/DelayedUpdate.h"
-#include "simd/simd.hpp"
+#include "CPU/SIMD/simd.hpp"
 
 #include <stdio.h>
 #include <string>
