+++ src/Sandbox/determinant_delayed_update.cpp
@@ -21,9 +21,9 @@
 #include <mpi/collectives.h>
 #include <getopt.h>
 using namespace std;
-#include <simd/allocator.hpp>
-#include <simd/simd.hpp>
-#include <simd/algorithm.hpp>
+#include <CPU/SIMD/aligned_allocator.hpp>
+#include <CPU/SIMD/simd.hpp>
+#include <CPU/SIMD/algorithm.hpp>
 #include <QMCWaveFunctions/Fermion/DiracMatrix.h>
 #include <QMCWaveFunctions/Fermion/DelayedUpdate.h>
 using namespace qmcplusplus;
